refix_DB

+++ db/migrate/20210601231546_create_goods.rb
@@ -1,8 +1,8 @@
 class CreateGoods < ActiveRecord::Migration[5.2]
   def change
     create_table :goods do |t|
-      t.integer :post_id, null: false
-      t.integer :user_id, null: false
+      t.bigint :post_id, null: false
+      t.bigint :user_id, null: false
       t.timestamps
     end
     add_foreign_key :goods, :posts
