+++ db/migrate/20210601231552_create_chats.rb
@@ -1,8 +1,8 @@
 class CreateChats < ActiveRecord::Migration[5.2]
   def change
     create_table :chats do |t|
-      t.integer :user_id, null: false
-      t.integer :room_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :room_id, null: false
       t.string :message
       t.timestamps
     end
